@@ -4,7 +4,7 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: '>=2.16.0 <3.0.0'
+  sdk: '>=2.17.0 <3.0.0'
 dependencies:
   cloud_firestore:
   cloud_functions:
@@ -20,6 +20,7 @@ dependencies:
   freezed_annotation:
   go_router:
   hooks_riverpod:
+  json_converter_helper:
   mono_kit:
   recase:
   rxdart:
@@ -31,7 +32,7 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
   freezed:
-  json_serializable:
+  json_serializable: '>=4.6.0'
   pedantic_mono:
 dependency_overrides:
   hooks_riverpod: '>=2.0.0-0'
