@@ -1,16 +1,17 @@
 name: firestore_ref
 description: Firestore reference/document typed access helpers. You can write organized safe code by this.
-version: 0.13.3
+version: 0.14.0
 homepage: https://github.com/mono0926/flutter_firestore_ref
 environment:
-  sdk: '>=2.13.0 <3.0.0'
+  sdk: '>=2.17.0 <3.0.0'
 dependencies:
   cloud_firestore: '>=3.0.0 <4.0.0'
   disposable_provider: '>=2.3.0 <3.0.0'
   firebase_core: '>=1.1.1 <2.0.0'
   flutter:
     sdk: flutter
-  json_annotation: '>=4.0.0 <5.0.0'
+  json_annotation: ^4.6.0
+  json_converter_helper: '>=0.0.1 <2.0.0'
   meta: '>=1.3.0 <2.0.0'
   quiver: '>=3.0.0 <4.0.0'
   rxdart: '>=0.27.0 <1.0.0'
