@@ -2,6 +2,7 @@ import 'package:firestore_ref/firestore_ref.dart';
 import 'package:flutter/foundation.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:json_converter_helper/json_converter_helper.dart';
 
 part 'user.freezed.dart';
 part 'user.g.dart';
@@ -10,10 +11,11 @@ final usersRefProvider = Provider((ref) => UsersRef());
 
 @freezed
 class User with _$User {
+  @allConverters
   const factory User({
     @Default(0) int count,
-    @TimestampConverter() DateTime? createdAt,
-    @TimestampConverter() DateTime? updatedAt,
+    @Default(UnionTimestamp.serverTimestamp()) UnionTimestamp createdAt,
+    @Default(UnionTimestamp.serverTimestamp()) UnionTimestamp updatedAt,
   }) = _User;
   factory User.fromJson(JsonMap json) => _$UserFromJson(json);
 
@@ -30,7 +32,11 @@ class UsersRef extends CollectionRef<User, UserDoc, UserRef> {
   UsersRef() : super(FirebaseFirestore.instance.collection('users'));
 
   @override
-  JsonMap encode(User data) => replacingTimestamp(json: data.toJson());
+  JsonMap encode(User data) => data
+      .copyWith(
+        updatedAt: const UnionTimestamp.serverTimestamp(),
+      )
+      .toJson();
 
   @override
   UserDoc decode(DocumentSnapshot<JsonMap> snapshot, UserRef docRef) {
@@ -49,10 +55,9 @@ class UsersRef extends CollectionRef<User, UserDoc, UserRef> {
 
 class UserRef extends DocumentRef<User, UserDoc> {
   const UserRef({
-    required DocumentReference<JsonMap> ref,
+    required super.ref,
     required this.usersRef,
   }) : super(
-          ref: ref,
           collectionRef: usersRef,
         );
 
