@@ -1,21 +1,27 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
+// ignore_for_file: type=lint, implicit_dynamic_parameter, implicit_dynamic_type, implicit_dynamic_method, strict_raw_type
+
 part of 'user.dart';
 
 // **************************************************************************
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$_User _$$_UserFromJson(Map<String, dynamic> json) => _$_User(
+_$_User _$$_UserFromJson(Map json) => _$_User(
       count: json['count'] as int? ?? 0,
-      createdAt:
-          const TimestampConverter().fromJson(json['createdAt'] as Timestamp?),
-      updatedAt:
-          const TimestampConverter().fromJson(json['updatedAt'] as Timestamp?),
+      createdAt: json['createdAt'] == null
+          ? const UnionTimestamp.serverTimestamp()
+          : const UnionTimestampConverter()
+              .fromJson(json['createdAt'] as Object),
+      updatedAt: json['updatedAt'] == null
+          ? const UnionTimestamp.serverTimestamp()
+          : const UnionTimestampConverter()
+              .fromJson(json['updatedAt'] as Object),
     );
 
 Map<String, dynamic> _$$_UserToJson(_$_User instance) => <String, dynamic>{
       'count': instance.count,
-      'createdAt': const TimestampConverter().toJson(instance.createdAt),
-      'updatedAt': const TimestampConverter().toJson(instance.updatedAt),
+      'createdAt': const UnionTimestampConverter().toJson(instance.createdAt),
+      'updatedAt': const UnionTimestampConverter().toJson(instance.updatedAt),
     };
